Finnish chapter four

+++ app/src/main/java/com/ivaldoalmada/activity/StudentListActivity.java
@@ -83,15 +83,18 @@ public class StudentListActivity extends AppCompatActivity {
     @Override
     public void onCreateContextMenu(ContextMenu menu, View v, ContextMenu.ContextMenuInfo menuInfo) {
         super.onCreateContextMenu(menu, v, menuInfo);
-        menu.add("Remove");
+        getMenuInflater().inflate(R.menu.activity_lista_alunos_menu, menu);
     }
 
     @Override
     public boolean onContextItemSelected(MenuItem item) {
-        AdapterView.AdapterContextMenuInfo menuInfo = (AdapterView.AdapterContextMenuInfo) item.getMenuInfo();
-        Student studentFound = studentListAdapter.getItem(menuInfo.position);
-        studentDao.remove(studentFound);
-        studentListAdapter.remove(studentFound);
+        int itemId = item.getItemId();
+        if(itemId == R.id.activity_lista_alunos_menu_remover) {
+            AdapterView.AdapterContextMenuInfo menuInfo = (AdapterView.AdapterContextMenuInfo) item.getMenuInfo();
+            Student studentFound = studentListAdapter.getItem(menuInfo.position);
+            studentDao.remove(studentFound);
+            studentListAdapter.remove(studentFound);
+        }
         return super.onContextItemSelected(item);
     }
 }
